requirements.txt
Django==5.1.2
djangorestframework==3.15.2
psycopg2-binary==2.9.9
pillow==10.4.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.5.0
python-dotenv==1.0.1
gunicorn==23.0.0
django-filter==24.3
drf-yasg==1.21.7
celery==5.4.0
channels-redis==4.2.0
boto3==1.35.39
django-celery-beat==2.7.0
django-celery-results==2.5.1
channels==4.1.0
daphne==4.1.2
requests==2.32.3
flower==2.0.1
Twisted[tls,http2]==24.7.0
aiohttp==3.10.10
pandas
scikit-learn==1.6.1
pickle